feat: Enable BuildConfig and use Long for version codes
Enables the BuildConfig feature in the app's build.gradle file. This allows access to generated constants such as version codes and names.

Additionally, modifies the UpdateManager to cast the BuildConfig.VERSION_CODE to a Long. This ensures accurate comparison with server-provided version codes, which can exceed the Integer limit for very large version numbers.

+++ app/src/main/java/com/ultra/optimize/x/utils/UpdateManager.kt
@@ -66,7 +66,7 @@ class UpdateManager(private val activity: Activity) {
                     val updateUrl = document.getString("update_url") ?: ""
                     val forceAutoUpdate = document.getBoolean("force_auto_update") ?: false
 
-                    val currentVersionCode = BuildConfig.VERSION_CODE
+                    val currentVersionCode = BuildConfig.VERSION_CODE.toLong()
                     Log.d(TAG, "Latest Version: $latestVersionCode, Current: $currentVersionCode")
                     Toast.makeText(activity, "Latest: $latestVersionCode, Current: $currentVersionCode", Toast.LENGTH_LONG).show()
